Update code.py
Insert new dataframe that will contain the result of all trades with the trailing stop and target
added the info for it to look for trade still on. 
still need to add a while loop that searches for the trade still on if true y must perform the same action as it did after the else:

After all this is done still need to reindex the Results dataframe if not the graph comes out wrong. It has been tried with Results.reindex()
and it did not work.

One important thing to keep in mind the len of the trade must be taken before this loops if not one trade with multiple days will look like multiple trades since each day count as 1

Short process must me done also